requirements.txt
django==4.1.3
django-environ==0.9.0
psycopg2-binary==2.9.5
django-mptt==0.14.0
pytz==2022.6
django-ajax-datatable==4.4.4